requirements.txt
numpy>=1.20.0
pyrtlsdr>=0.3.0
scipy>=1.7.0
matplotlib>=3.5.0
